requirements.txt
djangorestframework==3.13.1
python-dotenv==0.20.0
web3==5.28.0
